Since the introduction of 5p/3p miRNA annotation all miRNAs went into the mature.fa file in miRBase and deprecation of the star file. When using miRDeep2 the result.html has a field (column 5) indicating if the predicted mature sequence is in accordance with the miRBase annotation or if it is the star sequence of the miRBase annotation. This field should always be true when the sequence is found in the mature annotation file.
However, since there are now always 2 mature miRNA (5p and 3p corresponding to
the initial mature and star or star and mature sequence which have been by definition
the more abundant and the less abundant) sequences, also the shown 'mature miRBase miRNA' was always
the 5p sequence independent if this was the less abundant one in the prediction or not.

Please observe that the number of reads for the mature sequence and the mature sequence itself
were correctly given for the most abundantly detected sequece. E.g. If the 3p sequence was most
abundant then the counts and sequence were given as mature read counts and mature sequence (which is correct) but misleadingly the field 'mature miRBase miRNA' listed the 5p sequence.

This has been corrected now and only affects your result.html if you provide only a file
with 5p/3p annotation and no file with 'star' miRNAs.